fix: only write version file if supabase directory exists (#2408)

+++ cmd/root.go
@@ -153,7 +153,7 @@ func Execute() {
 func checkUpgrade(ctx context.Context, fsys afero.Fs) (string, error) {
 	if shouldFetchRelease(fsys) {
 		version, err := utils.GetLatestRelease(ctx)
-		if len(version) > 0 {
+		if exists, _ := afero.DirExists(fsys, utils.SupabaseDirPath); exists && len(version) > 0 {
 			err = utils.WriteFile(utils.CliVersionPath, []byte(version), fsys)
 		}
 		return version, err
